Rename "Version" and "rollback" to "Revision" and "revert"
"Version" is renamed to "Revision" for ease of use on the MongoDB shell, since
`db.version` is a reserved attribute.

From now on, the "revision" term will be used to represent the whole metedata
that's recorded for a given instance, and "version" for the content of the
"document" field (i.e. the `instance.__dict__`).

"rollback" is renamed because "revert" makes more sense.