Steal pymongo's `find_one` for the good cause
Currently if you want to modify the find behavior, you have to overload both
`find` and `find_one`.

The idea here is to keep the exact same behavior than pymongo's `find_one` but
also to be able to overload `find` in a Thingy with `find_one` that also
benefits from that overload.