Make the async test function detection more robust
The problem with "inspect.getsource" is that if the test function has
decorators, they will appear in the source code, making the condition invalid.

Using pytest internal function "get_real_func" allows us to access the original
function directly and check if the function is a coroutine much more robustly.